# mapaviajero-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Travel App API",
        "description": "Documentación de la API Travel App",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/marcaCoches": {
            "get": {
                "tags": [
                    "MarcasCoche"
                ],
                "summary": "Listar todas las marcas de coche",
                "description": "Obtiene la lista de todas las marcas de coche disponibles. Solo accesible para usuarios registrados",
                "operationId": "getMarcasCoche",
                "responses": {
                    "200": {
                        "description": "Lista de marcas de coche obtenida correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 3
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/MarcaCoche"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "MarcasCoche"
                ],
                "summary": "Crear una nueva marca de coche",
                "description": "Solo accesible para administradores y superadministradores. Permite crear y añadir una nueva marca de coches a la Base de Datos.",
                "operationId": "storeMarcaCoche",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MarcaCocheDB"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Marca de coche creada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/MarcaCoche"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/marcaCoches/{id}": {
            "get": {
                "tags": [
                    "MarcasCoche"
                ],
                "summary": "Mostrar una marca de coche específica (por id)",
                "description": "Obtiene los detalles de la marca de un coche por su ID. Solo accesible para usuarios registrados",
                "operationId": "getMarcaCocheById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la marca de coche",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Marca de coche encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/MarcaCoche"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Marca de coche no encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "MarcasCoche"
                ],
                "summary": "Actualizar una marca de coche",
                "description": "Permite modificar los datos de una marca de coches. Solo accesible para administradores y superadministradores.",
                "operationId": "updateMarcaCoche",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la marca de coche a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MarcaCocheDB"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Marca de coche actualizada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/MarcaCoche"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Marca de coche no encontrada"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "MarcasCoche"
                ],
                "summary": "Eliminar una marca de coche",
                "description": "Permite eliminar una marca de coches. Solo accesible para administradores y superadministradores.",
                "operationId": "deleteMarcaCoche",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la marca de coche a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Marca de coche borrada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Marca de coches borrada correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Solo el administrador puede borrar una marca de coches"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Marca de coche no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Marca de coches no encontrada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "MarcaCoche": {
                "title": "Marcas de Coche (Recurso API)",
                "required": [
                    "id",
                    "nombre"
                ],
                "properties": {
                    "id": {
                        "description": "ID único de la marca del coche",
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "description": "Nombre de la marca del coche",
                        "type": "string",
                        "example": "Toyota"
                    }
                },
                "type": "object"
            },
...
            "MarcaCocheDB": {
                "title": "Marca de Coche (BBDD)",
                "required": [
                    "nombre"
                ],
                "properties": {
                    "nombre": {
                        "description": "Nombre de la marca del coche",
                        "type": "string",
                        "example": "Seat"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}